Execute bq room available handler on p2p thread
Excecuting the block queue "room available" handler (i.e. the handler which is executed when the block queue's state goes from full -> not full) on the p2p thread rather than the client thread simplifies the threading story for BlockChainSync. Note that we still need to obtain the mutex in the handler because it modifies data (e.g. m_state) that's read on other threads (e.g. BlockChainSync::status() reads m_state and can be called on the Client thread).